tidewater: add initial consumer persistence setup
This puts an initial message processing place for the consumer that
expects a message payload to be sent of the form:

```json
{"resourceUrl": "https://tidewaterurl/to/resource/a", "status": "published"}
```

Where `status` may be any of: "published", "modified", "deleted"

A `Persister` class is created with some initial methods related to the
`status` commands.

Much of this is WIP, because we do not yet know the shape of the data.
So we are also not yet using the `resourceUrl` although we are raising
an error if it is not present, same with `status`