surfliner: add a document as the home for queing configuration
we need a documenation home for information about the queuing commitments of
various products.

this takes the opinion that configuration to talk to the queue should be
strictly OPTIONAL for each product, and then sets itself out as the place where
inter-product documentation about interactions with the queue should live.

i think we should put infor about our topics and routing keys here, e.g.

related to #953