tidewater: allow rabbitmq to determine queue name
This is probably beneficial for scaling.

Per the docs:

To ask an AMQP broker to generate a unique queue name for you, pass an
empty string as the queue name argument. A generated queue name (like
amq.gen-JZ46KgZEOZWg-pAScMhhig) will be assigned to the Bunny::Queue
instance that the method returns: