comet: add debug behavior for the Minio setup; allow skip with env
the docker-compose containers using the comet images sometimes get stuck in this
initialization stage. it's probably helpful not to try to setup this endpoint
for container runs where we know we won't use it. specifically, don't connect
when running db-migrate in docker.

do we also want to skip this in various init containers in k8s? i'm not sure the
helm charts have hooks for us to override there, but it might be a need.