tidewater: add optional rabbitmq service for dev
We're leveraging the `--profile` option which enables optional services
to be added to a compose file for local development.

We will likely rarely need to have a rabbitmq service enabled, but in
certain instances, such as testing the consumer script, it may be of
value.

We expose the RabbitMQ management endpoint/port so it can be accessed
via the browser if desired.