comet: track access from discovery platforms via Hyrax's ACLs
Hyrax has Web ACL style access control lists. this proposes tracking access to
resources from a given discovery system by using those ACLs.

when a resource becomes discoverable on a new platform, we add an ACL entry
granting mode `:discovery` (is this the right mode?) on a group associated with
the discovery platform. we may want to reconsider either the discovery mode, the
group structure (or consider using a specialized user).

the reasoning (such as it is) behind using Groups, is that it creates a
contextual model in which a user accessing a resource from, e.g. ~tidewater, is
authenticated to "discover" it by virtue of their present membership in the
tidewater group; this is similar to how Hyrax apps do IP based auth.

this seems good enough reasoning for now?

references #958;

this is an alternative model to the one proposed in !1966. if we move this
direction we may want to look at extracting other parts of that MR for
compatibility.

i worked on this with @kibigo, @mcritchlow, and @tchu1.