comet: document relationship to PCDM
document the fact of (if not a lot of detail about) Comet's use of PCDM.

i've tried to provide a bit of context about the intent of PCDM (though
admittedly, some of the goals were contentious at the time).

(`Hydra::Works` in particular has no clear definition as an abstract model, though
several definitions are written down or diagrammed. The real definition of the
`Hydra::Works` model is defined by the related library, which is itself
dependent on `ActiveFedora` implementation details... anyway...)