surfliner: install rabbitmq operator
We're going to try integrating the rabbitmq operator and installation
process into the CI/CD pipeline, in the context of Comet. We think Comet
makes sense as the entrypoint for ensuring that a RabbitMQ Operator (and
queue instance) is installed into the kubernetes cluster and namespace
in the current context (review, staging, production).

Co-authored-by: Matt Critchlow <mcritchlow@ucsd.edu>
Co-authored-by: Dan LaSusa <dlasusa@ucsd.edu>