comet: move discovery platform configuration behind an interface
remove duplicative access of RabbitMQ related environment variables, by
factoring them into an organized interface.

for now, this interface is just a Struct which accepts a `name` and constructs
an appropriate `#topic` and `#routing_key` for messages related to the
discovery platform. it's possible we'll want to expand this interface and/or
provide a mechanism for more complex configuration. a `Struct` might not serve,
in this case, but it seemed very low-cost to start out simply.

related to #957.