surfliner: setup surfliner_schema gem
This gem will serve as a hard fork of `Hyrax::SimpleSchemaLoader` for
Surfliner products (which may or may not have access to Hyrax).

I based the configuration of this gem on the existing one at
`gems/solr_utils` as well as samvera:valkyrie. At present the only
meaningful functionality is version number reporting; actual
schema-loading code TK.