comet: introduce a RabbitMQ wait process for dev and system tests
AMQP startup seems a bit hard to test (`db-wait.sh` doesn't seem to do the
trick), so this checks using the Ruby setup we use to actually connect. the
normal failure occurs at `connection.start`.

notably, we don't need this check in many scenarios. in Helm depolyments, we can
allow the app to fail on initialize and/or in health check (TBD) if the
connection isn't available. `CrashLoopBackoff` will wait for the service to come
up and we'll eventually get a healthy pod if the queue becomes available.

for system tests and the dev env though, we just don't want to start unless we
have a good connection; if we crash in these cases, we ususally won't come back
up.