comet: add an OkComputer status check for RabbitMq
i considered using this version of the status check:
https://github.com/sportngin/okcomputer/blob/master/lib/ok_computer/built_in_checks/rabbitmq_check.rb

but i think we aren't interested merely in whether we CAN connect with the
configuration, but in whether the current process's connection is actually open.
if we're persistently failing this check, i think we want to cycle the pod, not
hope that some other system will heal the connection.