comet: use a preloaded solr config
it looks like GitLab runner doesn't always checkout the repo before the solr
container entrypoint runs. make sure the config directory exists by preloading
it into the container using the gitlab runner configuration.

on the runner side, i added:

```toml
[[runners.kubernetes.volumes.config_map]]
  name = "comet-solrconfig"
  mount_path = "/surfliner-conf/comet/solr"
```

and loaded a config map into the runner namspace containing the whole of
`solrconfig.xml` and `schema.xml` from the current `comet` trunk.

if these files change, we'll need to MANUALLY maintain them. but this was the
best i could come up with to fix the runner situation.

issue #984 calls for follow-up
follow-up work is needed as documented in #984.