comet: connect to rabbitmq in review applications
Because the rabbitmq cluster instance is managed by the rabbitmq
operator, we expect this to be pre-existing and available for
interaction via the `Service`.

In order to uniquely publish messages for each comet review application,
we set a dynamic `topic` by appending a `$CI_MERGE_REQUEST_ID` suffix to
the topic name. This unfortunately is directly targeting the index of
the `extraEnvVars` array that contains the `RABBITMQ_TOPIC` environment
variable in `charts/snippets/comet/review-deploy.yaml`