superskunk: add METADATA_MODELS to helm chart
The addition of `extraEnvVars` is probably still a good one for the Helm
chart to allow as much flexibility as possible.

In the case of the `METADATA_MODELS` variable and usage, I think it's
integral enough for the application that I think it deserves its own
entry into the `values.yaml` and the chart `ConfigMap`.