shoreline: ignore contrib directory in pipelines
We are introducing a new `contrib` directory for holding scripts and
other tools that are useful for folks in the current Shoreline workflow,
but are not directly part of the Shoreline application itself.

Therefore, we don't need to run any of the existing CI/CD pipelines for
Shoreline if these files change.