superskunk: Initial ACL documentation
The exact details of this are TBD but here is an initial specification
of what we have been considering. It uses a two‐request model with a
nonce to prevent signature reuse, since I think that might be simpler
to implement than having nonces be generated on the consumer side and
remembered for the duration of their applicability. That is open to
discussion, though.

Undecided is how public key information gets *into* ~superskunk (maybe
via environment variables?), what algorithms we will support, and any
additional constraints we might need.

Also undocumented at this point is how exactly ~superskunk decides
whether to allow a consumer to access a file.