tidewater: initial keypair support
For #1000 we need to add support for the ~tidewater application to have
access to a keypair used for signing payloads sent to ~superskunk.

This adds a new `Secret` template to the Helm chart, which use of type
`kubernetes.io/ssh-auth` and holds both the public and private key for
the keypair. A default keypair is included in the template, but is
expected that in a `production` context a deployment instead will
provide a reference to an existing `Secret` via `keypair.existingSecret`

Documentation added to the Chart `README` for setting up the `Secret`

New Helm chart test added to ensure the keypair exists where it is
expected to be and can sign content. This test is currently only being
run for `review` applications.