surfliner: allow kaniko  to use --push-retry
Setting the value to `3` initially, which we can adjust as needed.

This feature was introduced relatively recently, the following PR

https://github.com/GoogleContainerTools/kaniko/pull/1578

Reason for feature:

> We are facing intermittent issues to push the image to the destination. Cause is as far as I can tell network flakeness. There is a long-standing issue asking for retries for the push operation, so I investigated this.

I recently noticed this error that happened to coincide with 5
dependency MRs from Renovate all coming in at once.

This user references similar experiences with Kaniko/Gitlab:

https://github.com/GoogleContainerTools/kaniko/issues/584#issuecomment-653047077

The hope here using the `--push-retry` option is that it will try to
push a few more times and avoid the need for us to manually retry the
`build` job to get it to pass.