starlight: Initial ucsb2 theme work (eyebrow)
Because the current ucsb theme is built off of an old Bootstrap 3
Drupal theme which it would be a pain to try to get working in
Bootstrap 4, I’m looking at just reimplementing at least the big stuff
like headers and footers from scratch.

This commit gives us the UCSB eyebrow and matches the appearance of the
one on <https://www.ucsb.edu/>.

Eventually this will replace the existing ucsb theme, but I’m working
on it separately for now.