WIP:comet:surfliner_schema: use a module builder for schemas
the goal of this work is to remove `SurflinerSchema::HyraxLoader` and use the
same interfaces for all schema application.

this WIP introduces `SurflinerSchema::Schema()` following the module builder
pattern used in Hyrax. this proposes droping the `:noop` schema names and using
the name to match on "availablity". that implementation isn't yet done (it
wasn't needed yet because Comet hasn't yet been matching on "availability").