starlight: ucsb2 footers
I’ve made some decisions regarding the responsiveness of the footer
elements which diverge a bit from the UCSB Drupal theme. I personally
hate it when mobile views give you a list of links a mile long at the
bottom of a page, which is a symptom of Bootstrap managing all of its
breakpoints based on viewport size with no attention paid to the actual
dimensions of content. We can do better with CSS Grid, so I have.

I’ll put some pictures in the MR so that people can review whether this
is acceptable, and if we need to go the Bootstrap way then we can. I
just worry it makes our (.html.erb) code much more complex and
difficult to maintain for negative benefit.