starlight: ucsb2 Site header & layout improvements
The major addition of this commit is the “UC Santa Barbara Library”
site header which now appears under the eyebrow on all pages.

I’ve also taken the opportunity to clean up some of the layout with
regards to max-widths and centring, and made use of the Bootstrap
`media-breakpoint-up` include for a couple of breakpoints which we need
to respect for Bootstrap reasons.

The “UC Santa Barbara Library” wordmark SVG included in this commit was
generated in Illustrator from the EPS file available through the
library wiki. I’m using this instead of the PNG files currently in‐use
by the `ucsb` theme because we may need to be able to choose its colour
at will (and also it just looks better).