starlight: ucsb2 theme exhibit mastheads
I had to make a number of somewhat arbitrary styling / responsiveness /
interactivity decisions with this, since the PDF I’m working from is a
static page at a fixed size with no calculations specified. I tried to
follow the look and behaviour of the main UCSB websites as much as
possible in this.