starlight: fail startup on translation initializer
If we are in a situation where we are NOT using the `nulldb` adapter,
then we should enforce the `Translation` table's existence. This can
happen when we're running the database migrations in a separate
container image, such as in local development and it has not finished by
the time this initializer runs. We would prefer to have the application
fail so it can be restarted and start properly once the table does
exist.