starlight: Allow theme‐specific localization
This commit monkeypatches the `translate` function available in views
to first try appending the current theme, in parentheses, to the
beginning of the localization key. So the key

    spotlight.header_links.login

becomes

    spotlight.header_links.(theme)login

These themed localization keys can be added as usual in the
localization YAML files. If no themed key is found, `translate` will
fall back to normal behaviour (i·e use the provided key without
modification).

This commit also adds two theme‐specific localizations for the `ucsb2`
theme.