starlight: add deviseGuestsCronJob for cleanup
The devise_guests gem, over time, creates a massive amount of "accounts"
that can generally be considered ephemeral. It does so for things like
visits, bookmarks, etc. The gem exposes a rake task for cleaning up
older account records called `delete_old_guest_users`.

In UCSB's case, this number was well over 2 million `2477495` and
causing issues with their deployment.

We are adding a k8s `CronJob` to run this rake task in a given
`schedule` which can be set as desired by using typical Cron syntax.