Starlight: Let themes add to site sidebar
If a file `app/views/shared/starlight/_sidebar_preamble_$THEME` exists
for the current theme, then it will be rendered into the site sidebar
before the existing administrative sidebar links.

This check is done via `ActionView::LookupContext::ViewPaths#exists?`,
which is not well‐documented but seems to do the trick.