starlight: use fixed path for UV
We had previously used a md5 hashed path for the location to serve
universal viewer out of. Unfortunately, since this changes quite
regularly, it would create a situation where an embedded item would like
to an image which would no longer resolve after a new deployment.

This sets the path at a fixed `/uv` to avoid this issue going forward.

This change may result in some cache invalidation issues when Universal
Viewer is upgraded, however we do so quite infrequently.