starlight: upgrade omniauth to version 2
OmniAuth 2 recommends using `post` instead of `get` methods for sign in.
This makes sense, but would require updating some upstream
authentication partials to support passing in a `method` to the various
`link_to` calls for signing in, so we could specify `post`. It would
also require rewriting some of our tests and authentication code.

For these reasons, we first want to preserve existing authentication
behavior but also upgrade these dependencies. And tackle the `post`
issue as a follow up.