starlight: add_exhibit_admin rake task added
Since there is an issue with adding exhibit administrators via the UI
currently, it would be helpful to have a rake task to do so at the
command line.

The `spotlight:admin` rake task exists but adds the user as a sitewide
admin, which we often do not want.

This task prompts the user for:
- exhibit slug
- email address of user to add as exhibit admin