surfliner_schema: Refactor Loader to be Hyraxish
Loader should now be usable with the SurflinerSchema module builder
pattern, instead of exposing readers which applications will manually
have to handle. This winds up producing quite a bit of A·P·I overlap
with ⸺Reader, as things which are available on ⸺Readers also
need to then be exposed on a Loader somehow. The most significant
difference between the two is that ⸺Readers are file‐ and
format‐specific, and Loaders can contain multiple ⸺Readers.