surfliner_schema: Initial Houndstooth support
This commit just enables reading Houndstooth files into the existing
schema data structure. It does *not* add support for validations or
any of the more advanced M3 features we have been discussing; the goal
is just to enable UCSB to define our schemas using the Houndstooth/M3
model with the same level of support currently given to Hyrax Simple
Schemas.