starlight: support creation of ServiceAccount
This is included by default in new charts, but our chart is old enough
to not have this template/setup.

Having a `ServiceAccount` setup is helpful for security reasons.
Specifically, in order to be able to leverage Hashicorp Vault external
secrets, a `ServiceAccount` must be used in setting up the `role` that
allows the application to request secrets from the Vault.