starlight: update jobs to support vault secrets
Rather than cherry-pick variables and map them using `env`, we can pass
the `ConfigMap` and `Secret` values to the `util` image instead.

This does require mapping some values from how they're named for the
chart, or specific Bitnami images, to the way the `aws` and
`postgresql-client` utils need the env vars to be.