starlight:ucsb: use the existing secret for solr authentication
this secret was incorrectly indented, causing SOLR to have the wrong password in
its environment variable.

it appears that the actual password is correct (since it has been set since
before this indentation became necessary in
https://github.com/bitnami/charts/commit/205818407e7517de56514f2bb37cc1b9217ef8a8.
but with the move to chart version 4.0.9+, Solr uses the environment variable
set by this secret to authenticate to itself for liveness/readiness
probes. those probes have been failing on that chart version in this build
because the password variable (but NOT the password) has been set to the default
"admin".