starlight: ucsb: .active styling 🔜 navbar links
We only do the thick bottom border thingy in the horizontal (wide)
layout because it really doesn’t make sense when the links are being
stacked vertically (i·e on narrow screens). In wide/horizontal layout,
hovered and active links now have the same appearance (matching UCSB
branding elsewhere).

In the narrow/vertical layout, I’ve made the .active link bold and the
hovered link (was already) underlined. This is probably the approach we
will wind up with for vertical link lists elsewhere on the site, for
example feature pages.

With respect to the vertical layout, this commit also takes the liberty
of reducing the horizontal padding of the navbar search form because it
was really irking me.