starlight: ucsb: Make about sidebar links 2ⁿᵈʳʸ
The entire contents of the about sidebar was being styled like a
header due to about sidebars not being nested in the same way as
feature page ones.

This adds an additional test to check if .sidenav’s are being preceded
by a .nav-heading. If they are, their list items are styled as if they
were nested an additional level.