starlight: artifact solr config for CI test job
We are changing the setup for Solr configset association with a Solr
core to leverage the Gitlab CI artifacts feature. Unfortunately, we
don't get direct access to the artifact in `service` images, so we have
to use the Gitlab REST API to get the artifact instead. After doing this
and placing it in the SOLR_CORE_CONF_DIR location, we can invoke the
original `ENTRYPOINT` and `CMD` for the Solr image.

Co-authored-by: Kristian De Castro <kddecastro@ucsd.edu>