superskunk:comet: use UCSB houndstooth metadata for UCSB deploys
both superskunk and comet should use the UCSB metadata for UCSB deploys.

the current houndstooth file is still fairly basic, but we should be able to
replace it with a full scale configuration quickly if the baseline works. this
is in service of moving forward with a real deployment of UCSB metadata.