comet: remove noop user from REDIS_URL
Not providing a username seems to use the default user along with the
`REDIS_PASSWORD`, which seems like what we want for dev/test/ci.

We were getting the following error:

```sh
The Redis connection was configured with username "noop", but the
provided password was for the default user. This will start failing in
redis-rb 5.0.0.
```