surfliner_schema: Update Schema language
Hyrax thinks of schemas in terms of “name”, but this gem mostly thinks
of them in terms of “availability”. I think the latter is more
consistent with current M3 and avoids some confusion with schema
mappings (which also have names).

The precise meaning of “availability” is left to be determined, but in
practice it is whatever is set by `available_on.class` in the M3 model.