comet: make component test resilient to browser timing
pulling the component id from the current page creates a race condition which
results in this test sometimes failing looking for "new" (from the URL) in the
parent's members.

this newer test might(?) still have timing issues, but at least it will fail
with an obvious error: "no, there are no members". if that's an issue, we can
add a wait later on.

since we're not testing the creation of the parent, let's create it out of
band.