ci: silence kubeconfig warnings
In CI deploy jobs using the `KUBE_ROLE_CONFIG` variable to setup a
`KUBECONFIG` file, warnings are generated in logs such as:

```sh
WARNING: Kubernetes configuration file is group-readable. This is insecure. Location: /etc/deploy/config
WARNING: Kubernetes configuration file is world-readable. This is insecure. Location: /etc/deploy/config
```

So we're ensuring the `KUBECONFIG` file is not group or world readable
via `chmod 600`