comet: support download for derivative content
allow users to download files within a fileset by their type.

the given type is mapped to a URI using the Hyrax::FileMetadata::Use module and
checked against the file's `type` field.

this fixes a misalignment where we previously relied on FileMetadata, the file,
and FileSet to have a very specific set of ids. the model supported by Hyrax is:

  FileSet.file_ids -> files in storage adapter <- FileMetadata.file_identifier

we use the custom query supported by Hyrax to handle resolution of FileSet to
FileMetadata. prior work implemented a version of this query for our metadata
adapter; we may want to refine that for efficency later.