superskunk: Set up controller to use OAI mappings
This shouldn’t have any effect right now, as the controller doesn’t
actually use a real model. That means there also aren’t tests; once
the ACL work is in and ~superskunk can read real models we should
fill out some tests for this behaviour.

Right now the `title` attribute is special‐cased as it is provided by
Hyrax rather than SurflinerSchema. I’d like to change this at some
point in the future but I think it is fine for now.

This code defines a `:mappings_for` private method on the controller
to make things easier should we decide we want to change how mappings
are provided on the model. It also assumes a correspondence between
the mapping iri defined in the schema and the profile requested in the
`Accept:` header (they should be the same).