comet: invert handling of backgound job for characterization
Comet should expect `object.file.uploaded` to be published every time a file is
uploaded. since either us or Hyrax are supposed to do this inline, we shouldn't
count on this event to only be published in a background job.

instead, publish each time we upload and override the appropraite listener to
enqueue a background job. the current setup assumes we want to characterize
every file we touch, but index the parent FileSet only when the
pcdm:OriginalFile's metadata is changed.