comet: Test metadata schema loading on objects
This commit adds a spec fixture with some basic M3 metadata and a test
to ensure that the metadata is properly loaded. The test is a bit messy
as it reads the properties directly from the YAML file instead of
hardcoding specific properties to check—I felt this was easier to
maintain.