comet: Use `.` rather than `:` in platform group
It would be easiest for ~superskunk if the ACL group name were a valid
HTTP token, so that it could be sent along with requests in e·g a
`User-Agent:` header. The allowed non‐alphanumeric characters in HTTP
tokens are: ``!#$%&'*+-.^_`|~``. I’ve chosen a period, but would be
open to other solutions, for example a `|`.

Otherwise, ~superskunk would need to maintain a list of
HTTP token–to–group name mappings, which seems cumbersome to me.