superskunk: Add DiscoveryPlatform
Right now this isn’t used, but it is designed to work in conjunction
with ACLs once the hyrax-acl gem is added.

~superskunk now requires a `User-Agent` HTTP header, which it uses to
determine the discovery platform name (by just taking the first
product name and ignoring everything else). This header will need to be
present in ~tidewater before the two will be able to properly send
messages to each other.

Once we have actual auth in place, the mechanism for determining the
discovery platform will probably change.