superskunk: Refactor ResourcesController
This eliminates the use of `public_send` in the controller and extracts
much of the behaviour out into dedicated serializer objects. The new
pattern is to define serializers inheriting from `ResourceSerializer`
and then to call `ResourceSerializer.serialize` to get the JSON‐LD for
a given resource with a given profile.