superskunk: Turn on reading from database
Superskunk now queries the comet database for resources. Actually
returning their metadata is pending the implementation of
`DiscoveryPlatform#has_access?`, so this will result in a 404 outside
of test environments.

An override has been added to the configuration to always allow access
to a platform with the name `"surfliner.superskunk.test_suite"` in the
testing environment *only*.

The OAI/DC test has been expanded to ensure that the returned JSON
actually contains the expected metadata.