surfliner: use dld tags for test jobs
At the moment, the ucsd k8s runner doesn't appear to be able to handle
running test jobs with a variety of services. So until we can properly
sort this out, and have a shared artifact hosting solution, we're
limiting jobs to only run on the dld k8s runner.