comet: disable RABBITMQ_ENABLED in staging and prod
neither campus appears to have the necessary configuration to turn this
on. we'll need to sort that out.

until then, disable just on these deploys; the review app deploys are still
fine, and we do want to keep the on-by-default behavior. this is just to keep
the build green while we fix the status-quo.