comet: add a rake task to delete an exchange
it's possible in rabbitmq for exchanges/topics to get created with the wrong
configuration. if the configuration isn't correct and aligned across publishers
and subscribers, and nothing critical has yet been published, the best thing to
do is to delete the exchange and start fresh with the right configuration.

this introduces a rake tasks for cleaning up exchanges in this case.