surfliner_schema: Add FormFields module builder
We want the metadata fields in the “New Object” page in ~comet to have
display labels and usage guidelines derived from their M3 property
definitions. This means that we need a way to determine, from a given
form :—

1. Whether a given property is defined in the M3 model or elsewhere,
   and furthermore

2. The values of various aspects of its definition (:display_label
   etcetera)

We *could* hack this in by just doing some creative calls to
`::EnvSchemaLoader.properties_for`, but this felt like a lot of weight
for a view. This commit attempts to solve the problem instead by
porting/extending `Hyrax::FormFields`
(now `SurflinerSchema::FormFields`) to supply an additional
`#form_definition` field, which can be called to get the definition for
any `SurflinerSchema`‐provided field available on a form, and which
returns `nil` if the field was not created by `SurflinerSchema`.

The tests are a little creative because I didn’t want to add
`Reform::Form` as a dependency (I just mocked `property` and
`validation` instead).

`SurflinerSchema::FormDefinition`s are `Delegator`s to
`SurflinerSchema::Property`s, which isn’t normally a pattern I would use
but seemed appropriate in this case.