comet: fix Hyrax's FindFileMetadata query
Hyrax's default FindFileMetadata query has some out of date/Wings specific
assumptions about File-FileMetadata-FileSet structure. we already use a custom
query for our real Sequel adapter, so we just need a replacement for the
in-memory test adapter.

this isn't an efficient query, but should be fine for the test suite. it's
substantially taken from a version written by @lsitu in the comments to !2561.