superskunk: Don’t let bad headers override errors
Superskunk’s error handling uses the `Accept` header to determine if
the error response should be plaintext or JSON. Formerly, if the
`Accept` header was bad, this would override the intended error with
a 400.

This commit rescues `AcceptReader::BadAcceptError` inside the error
handler to just display the plaintext version of the error in this case
instead.