starlight: universal viewer -> openseadragon
Several areas of the Spotlight codebase, such as the Item Embed widget,
already rely on openseadragon for the IIIF image viewer. Let's use it
for the main catalog show view as well, removing the need for yarn and
managing UV config and setup.

Note that the openseadragon partial itself comes from the
blacklight-gallery gem.