comet: don't create new Projects passively at runtime
Hyrax calls this `.find_or_create_default_admin_set` method regularly, and
expects it to find it by a static id, provided in a constant. since we don't
have a static id, and the concept of a "Default Project" is a bit dubious to
begin with we override.

The result of Hyrax's calls (from the ResourceForm, etc...) has been that a
default project is getting created over and over again, each one with a
different id, but with the same title. The overrides avoid this.

We do one potentially expensive lookup, then cache the result. If there's not an
existing Project called "Default Project", just assume there's no such concept
in the implementation and use whatever project pops up first. Count on user
input to determine the correct Project.