starlight: get create_exhibit_spec passing again
suddenly the webmock response did not include the `uri` and broke this
spec.

the behavior experienced in the test is the same as described in this
issue: https://github.com/bblimke/webmock/issues/469

It appears that passing the `net_http_connect_on_start` option to
Webmock, along with not requiring ssrf protection via
`skip_ssrf_protection`, allows the test to properly process/ingest the
sample image and pass the tests again.