comet: set systemwide `collection_model` to the Valkyrie model
set the system `collection_model` configuration to `Hyrax::PcdmCollection`.

because we index collections as `has_model_ssim` `"Collection"` we need to
override `SolrDocument#collection?` to ensure solr documents know when they
represent a Collection. the implementation included here is a bit of a catch
all, which should make things easier for us if/when we change the indexing
behavior. I think we should be able to do make a chang, but even if we change it
now, we'd need to reindex before removing the catch-all behavior.