surfliner_schema: Add fixture for property groups
Trying to namespace this into some kind of extension field felt like it
made the file much harder to maintain, so I just added the properties
directly (although this opens us up to conflicts if upstream Houndstooth
ever adds a different `sections` or `groupings` feature).

The syntax in this commit makes grouping membership inside of sections
implicit :—

- `date_uploaded` is in the section `basic_metadata` and the grouping
  `date`
- Therefore, `basic_metadata` must contain a `date` grouping

This is to save the trouble of having to remember rules like “specify
the section only on properties without a grouping” or trying to figure
out how to resolve cases where the section of a grouping and the
section of a property within that grouping don’t match.