surfliner_schema: Let loaders generate models
This commit introduces a new `resource_class_resolver` method onto
`SurflinerSchema::Loader` which returns a proc that can dynamically
create new model classes (and define them as global consts) for those
classes in the M3 schema.

Already‐defined consts take priority, so we are still able to manually
define classes for things like `Collection`. The base class used for
model definitions is configurable through subclassing or by explicitly
passing a base class to the `resource_class_resolver` method.

There is some flexibility with respect to the mapping from M3 into the
generated class name; both class names of the form `my_class` and
`MyClass` are supported in the M3, and will resolve to a Valkyrie
`internal_resource` of `MyClass`. To phrase it the other way, when
attempting to resolve `MyClass`, this code first checks for a `MyClass`
definition in the M3, and if it doesn’t find one falls back to
`my_class`.

Classes which aren’t defined in the M3 are *not* dynamically generated
and will continue to throw errors upon resolving.

This commit nominally introduces a dependency on active_support, but
that’s okay because Valkyrie already depended on it!