surfliner_schema: Set internal_resource for models
Dynamically‐created classes don’t have an appropriate value for
`to_s` at the time at which the default value for `internal_resource`
is defined in Valkyrie, so it is necessary to override the value in an
initializer.

An alternate approach might be to redefine the Dry type of the
`internal_resource` attribute to provide a different default.