comet: Use a generic ResourcesController
It looks as though simply declaring all the M3 models and associated
controllers up‐front in the Hyrax initializer is going to be the
easiest way to go about this. The routes need to be specified at
application startup, and each controller needs to be its own subclass
in order to support the `.curation_concern_type` static method. This
does mean potentially less work revising tests, as those declared
with constant forms of `GenericObject` ⁊·c will continue to work.