surfliner_schema: Support section and groupings
This commit introduces a new generic “Division” model which wraps (as
a Delegator) an underlying class, section, or grouping descriptor and
makes available its items, accessible from loaders via a new
`#class_division_for` method. I’ve designed this as a kind of tree
structure to hopefully be easy to iterate over in Hyrax presenter
logic, and to enforce a separation between *descriptors* (`Property`,
`ResourceClass`, `Section`, `Grouping`, which are boring structs) and
the more complicated object types necessary for this kind of work.

Hopefully this is good. I think the entire ~surfliner_schema test suite
probably needs a refactor, but I got in some basic tests with the
fixtures we have. There is probably a little bit of work needed to fix
up ~comet and ~superskunk to accommodate the changes I made to classes
here, which I will add in the following commits.