superskunk: Formalize Aardvark serialization
This commit removes support for the temporary
`tag:surfliner.gitlab.io,2022:api/shoreline/ingest` profile and replaces
it with a new `tag:surfliner.gitlab.io,2022:api/aardvark` one. This
tries hard to always produce well‐formed Aardvark JSON (plus a JSON‐LD
`@context`), but this isn’t to say that the resulting JSON is
necessarily *complete* in the sense of having everything which is needed
for ingest.

There are a few details here which probably merit further discussion :—

- Right now we are using a namespace of
  `http://geoblacklight.org/schema/aardvark/` for `gbl_`‐prefixed terms
  in Aardvark. This matches what is currently in the UCSB M3 file, but
  I’m not sure that this URI has been standardized. I at least couldn’t
  find it documented anywhere on the Open GeoMetadata page.

- This returns the Hyrax “last modified” date for the value of
  `gbl_mdModified_dt`. Is this what we want? Do we need to do more work
  to format this correctly (beyond just a `.to_s`)?

- The value of `dct_accessRights_s` defaults to `"Public"`, which
  we agreed was sufficient until we actually support restricted access.

- The value of `gbl_resourceClass_sm` defaults to `"Dataset"`. Is this
  right? How should this value be determined?

- Right now this assumes a straightforward mapping from whatever is
  stored in Valkyrie to OGM Aardvark. I’m not sure this is actually true
  (e·g for things like “Date Range”, how is that actually going to be
  represented in Comet / stored in Valkyrie? Probably not as a Solr date
  range string, right?). So, the metadata produced by this MR is “valid”
  in the sense that it meets the obligation, cardinality, and field type
  specified by Aardvark, but it is not necessarily “useful” or conformant
  to the Entry Guidelines.

Having this actually return meaningful data depends on having M3 schemas
which define mappings for the Aardvark URIs as defined in
`AardvarkSerializer::AARDVARK_TERMS`, and having that mapping be
associated with the `tag:surfliner.gitlab.io,2022:api/aardvark` IRI.